chore: add unit tests
Includes initial unit test setup and configuration, which was a little painful, especially when configuring with the ViteJS SWC compiler. But it was worth the effort to configure, as the unit tests runs pretty fast using the `@swc/jest` transformer module for tsx test files, and would have a positive net effect on time and cost on CI platforms. Fast 😎